Add GCC static analysis branch
Add a daily build of the git-only branch "dmalcolm/analyzer" hosted on
the GCC git mirror.

See https://gcc.gnu.org/wiki/DavidMalcolm/StaticAnalyzer

The branch adds support for in-tree GCC plugins, adding a
  --enable-plugins=
configuration option analogous to --enable-languages, and
the analyzer is itself implemented as the first example of
such an in-tree plugin.

Hence add a PLUGINS variable to build.sh analogous to LANGUAGES.