Add the hdf5 library to conan builder (#695)
* Add the hdf5 library to conan builder

This adds into the libraries configuration the hdf5 library.
The hdf5 library can be cloned from github. The configuration checks out
stable (1.12.1) and latest experimental (1.13.1) releases and configures
the cmake builds with everything but shared library building off.

* Add after_staging script to make sure include files are installed

hdf5 does not provide its public headers in an `include` dir. They are
placed in the installation target `include`. I am reusing the same
method to generate the expected include dir as found elsewhere in the
libraries.yaml.