Add the jakt programming language (#765)
Jakt does not currently (and for the forseeable future) provide any
precompiled binaries. So the compiler is build in misc-builder.

There are two compilers, a rust-based one that's build using cargo, and
then a selfhosted compiler that's build using the rust-based compiler
(for now).

This installs two compilers as one, since we need our rust-based
compiler to compile the selfhosted compiler.