Use hip-amd
The AMD implementation of the HIP interface is just called hipamd by
AMD. While hip-runtime-amd would probably be a reasonable name, it's
probably better to stick closer to what AMD is calling it.

Another reason is that that HIP variant for the NVIDIA platform is
header-only. That is, there's no runtime component. This naming will
therefore make a future hip-nvidia package a bit more consistent.