Add ROCm HIP library (#834)
* Add ROCm HIP library

* Remove use of install_always

The install_always option will cause the library to be installed again
even if already installed. That's not useful for release versions.

* Use the paths from the fixed misc-builder

It appears that this would be the default untar_dir if the target were
rocm-4.5.2, rocm-5.0.2, etc. However, since the target does not include
the rocm prefix, it needs to be specified.

* Use hip-amd

The AMD implementation of the HIP interface is just called hipamd by
AMD. While hip-runtime-amd would probably be a reasonable name, it's
probably better to stick closer to what AMD is calling it.

Another reason is that that HIP variant for the NVIDIA platform is
header-only. That is, there's no runtime component. This naming will
therefore make a future hip-nvidia package a bit more consistent.

* Deduplicate hip-amd