Replace baksmali check_file with check_exe
Adds a dependency on JDK 16.0.1 so that the version of the baksmali jar
can be checked with `java -jar <jar> --version` in check_exe instead of
just checking for the existence of the baksmali jar file.

The method for invoking the installed java is brittle, as it relies on
the previously-installed java path relative to the current working
directory (staging).