Update LLVM Flang file names
The latest builds contain:
```
$ ./bin/flang-
flang-new             flang-to-external-fc
```

The project has shifted the names over time.

flang was a script that called out to gfortran, and flang-new
was a compiler binary.

Then the flang script was renamed to flang-to-external-fc, but
its purpose remained the same.

The project has been considering renaming flang-new to simply
flang, but this has not happened yet.