Remove customization of Unifex's config.hpp
As of Unifex's PR 606[1] this past April, Unifex's `config.hpp` is
self-contained and does build-time configuration with the preprocessor
so we don't need a custom configuration step anymore.

This PR deletes the Compiler Explorer's copy of the customized
`config.hpp` and the build step that overwrites the upstream header with
that custom copy. I expect this change to resolve #6711 but I don't know
how to test that claim.

1. https://github.com/facebookexperimental/libunifex/pull/606