Add attempt to enable cpu delegation, plus debugging
We have an apparent race condition at startup where cgroups
aren't properly set up before we try and create. The error is
that nsjail fails due to lack of CPU delegation, even though the
cgcreates all succeed earlier on. We're not sure what's going on,
this is an attempt to "fix" plus diagnostics.