Enable CPU accounting - cgroup fix
There's a window during startup where, if we haven't started using the
cgroup controller with "cpu" in it, systemd may reload its configuration
and determine that nothing needs it, taking it away from us. Subsequent
`nsjail`s will fail with error 255.

Seemingly once `nsjail` has run at all, the `cgroup`s (even though not in
use) make the CPU set "busy" (we can't remove cpu from `/sys/fs/cgroup/cgroup.subtree_control`
gettin EBUSY). Presumably this is what "protects" running instances from
losing cpu control once we've booted up. Systemd reloads but can't remove the (apparently
unused) cpu cgroup control.

The correct fix is to make the compiler explorer service declare that it needs CPU
control. That seems to be `CPUAccounting=yes` *AND* `CPUQuota=100%` (see
https://unix.stackexchange.com/a/757423/167092).

I was able to reproduce the "startup race" in an isolated staging instance, and
show that adding these two lines fixes the issue: in fact, now systemd itself
adds the "cpu" so in theory we will be able to remove the code that adds it
in start-support.sh

See #1761 - we believe this fixes it